+++ python/libcasm/irreps/__init__.py
@@ -3,6 +3,7 @@ from ._irreps import (
     IrrepDecomposition,
     IrrepInfo,
     IrrepWedge,
+    MatrixRepGroup,
     SubWedge,
     VectorSpaceSymReport,
 )